requirements.txt
python-telegram-bot==21.5
anthropic==0.25.0
gspread==6.1.2
google-auth==2.29.0
httpx==0.27.2
